Add files via upload
Part1: 
For each of the following steps of analysis, choose an appropriate type of plot for comparing the data. Also, ensure that the plots are in increasing or decreasing order for better comparison. For example, if you make a bar plot to compare the GDP's of the states, ensure that the bars are in either increasing or decreasing order of GDP.
Part2:
In this part of the analysis, you will investigate whether there is any relationship between per capita GDP with dropout rates in education.